Refactoring to have more unit tests with standard input/output files
