+++ package.json
@@ -36,6 +36,7 @@
   },
   "scripts": {
     "build": "npx tsc",
+    "clean": "rm -rf ./build",
     "lint": "npx eslint \"**/*.ts\" \"**/*.js\"",
     "test": "./node_modules/jasmine-xml-reporter/bin/jasmine.js ./build/__tests__/*.test.js --junitreport --output=build/"
   },
