+++ tests/cases/commas-variables.glsl
