+++ tests/cases/commas-variables.json
@@ -1,0 +1,5 @@
+{
+  "sourceCode": "void main(){float A,B,C=0.5,D=1.;A=0.5,C=0.5;gl_FragColor=vec4(A,B,C,D);}",
+  "consts": {},
+  "uniforms": {}
+}
