+++ tests/cases/comments.glsl
