+++ tests/cases/comments.json
@@ -1,0 +1,5 @@
+{
+  "sourceCode": "void main(){}",
+  "consts": {},
+  "uniforms": {}
+}
