+++ tests/cases/const.glsl
