+++ tests/cases/define.glsl
