+++ tests/cases/define.json
@@ -1,0 +1,5 @@
+{
+  "sourceCode": "",
+  "consts": {},
+  "uniforms": {}
+}
