+++ tests/cases/hello.json
@@ -1,0 +1,5 @@
+{
+  "sourceCode": "void main(){gl_FragColor=vec4(1.,1.,1.,1.);}",
+  "consts": {},
+  "uniforms": {}
+}
