+++ tests/cases/include1.glsl
@@ -1,0 +1,1 @@
+@include "../cases/comments.glsl"     // Add some spaces and a C++ comment for good measure
