+++ tests/cases/include1.json
@@ -1,0 +1,5 @@
+{
+  "sourceCode": "void main(){}",
+  "consts": {},
+  "uniforms": {}
+}
