+++ tests/cases/include2.glsl
@@ -1,0 +1,2 @@
+uniform vec2 myUniform;
+@include "comments.glsl"
