--- tests/glsl/commas-uniforms.min.glsl
@@ -1,1 +1,0 @@
-uniform float A,B,C;void main(){gl_FragColor=vec4(A,B,C,1.);}
