--- tests/glsl/include.glsl
@@ -1,1 +1,0 @@
-@include "../glsl/comments.glsl"     // Add some spaces and a C++ comment for good measure
